maximum Likelihood update. Added parameters to describe the positionnal error of atoms in the model, in each ScatteringPowerAtom. These isotropic errors are used to compute Luzzati factors and derive the 'most likely' structure factors in ScatteringData, with a variance associated to each calculated structure factor. This is used in PowderPattern to calculate the 'most likely' pattern, and associate a variance to each caculated point of the pattern, and therefore derive the overall variance (cumulating observed and model error) and the associated weight.
Note that the above includes severe approximations to maximum likelihood pronciples,
to keep pure gaussian likelihoods which are fast to calculate. More strict calculations
of joint probabilities would require slower computations.

This also includes several optimizations, to compute directly 'integrated'
powder patterns (and the associated variances), without explicitely computing
the full patterns. This leads to a computation speed close to that of extracted
structure factors.

Also a bit optimized is ScatteringData::PrepareCalcStructFactor(), using clocks.
The function is still as ugly as before, though.